\name{builder.cacheSweave}
\alias{builder.cacheSweave}
\title{This function will build Sweave (Rnw) files using the cacheSweave driver.}
\usage{
  builder.cacheSweave(file, theenv, ...)
}
\arguments{
  \item{file}{the source file to build.}

  \item{theenv}{the environment to build in.}

  \item{...}{other unspecified parameters}
}
\value{
  the name of the file if successfully built.
}
\description{
  This function will build Sweave (Rnw) files using the
  cacheSweave driver.
}

