\name{buildVersion}
\alias{buildVersion}
\title{This is an internal method and should not be called directly.}
\usage{
  buildVersion(pv, version.major = NULL, ...)
}
\arguments{
  \item{pv}{the Project}

  \item{version.major}{the version that should be built.
  This can be either the the major number or version name.}

  \item{...}{other non-specified parameters}
}
\description{
  Builds a new version of a project. See Project$build
}

