\name{checkProject}
\alias{checkProject}
\title{This internal method will check to see if the Project class is current with
respect to the PROJECT.xml file. If it is out-of-date it will re-read the XML
file and return a new Project object.}
\usage{
  checkProject(pv)
}
\arguments{
  \item{pv}{the Project.}
}
\description{
  This internal method will check to see if the Project
  class is current with respect to the PROJECT.xml file. If
  it is out-of-date it will re-read the XML file and return
  a new Project object.
}

