\name{env2string}
\alias{env2string}
\title{This function will convert the elements in the given environment to a single
character string. This is useful when passing the contents of the environment
to a new R instances vis-a-vis the command line.}
\usage{
  env2string(theenv)
}
\arguments{
  \item{theenv}{the environment to convert to a string.}
}
\value{
  the string representation of the environment.
}
\description{
  This function will convert the elements in the given
  environment to a single character string. This is useful
  when passing the contents of the environment to a new R
  instances vis-a-vis the command line.
}

