% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{find_avg_stage_time}
\alias{find_avg_stage_time}
\title{Find Average Stage Time for an Item}
\usage{
find_avg_stage_time(item_data, stage_time_postfix = "_Hours")
}
\arguments{
\item{item_data}{A tibble containing data for a specific item.}

\item{stage_time_postfix}{A character string indicating the postfix
used to identify stage time columns.
Default is "_Hours".}
}
\value{
A tibble with the average time spent in each production stage
for the specified item.
}
\description{
Calculate the average time spent in each production stage
for a specific item.
}
\examples{
find_avg_stage_time(sample) # Calculate average stage times for all items in the sample dataset
}
