% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{find_stage_quantities}
\alias{find_stage_quantities}
\title{Find Stage Quantities for an Item}
\usage{
find_stage_quantities(
  item_data,
  stage_col = "Current_Stage",
  qty_col = "Quantity"
)
}
\arguments{
\item{item_data}{A tibble containing data for any number of items.}

\item{stage_col}{A character string indicating the column name for the current production stage.
Default is "Current_Stage".}

\item{qty_col}{A character string indicating the column name for the quantity of items.
Default is "Quantity".}
}
\value{
A tibble with the total quantity of items at each production stage.
}
\description{
Calculate the total quantity of items at each production stage.
}
\examples{
find_stage_quantities(sample) # Calculate stage quantities for all items in the sample dataset
}
