% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{find_total_production_time}
\alias{find_total_production_time}
\title{Find Total Production Time for Items in given data set}
\usage{
find_total_production_time(item_data, stage_time_postfix = "_Hours")
}
\arguments{
\item{item_data}{A tibble containing data for any number of items.}

\item{stage_time_postfix}{A character string indicating the postfix used to identify stage time columns.
Default is "_Hours".}
}
\value{
A tibble with the total production time for each item.
}
\description{
Calculate the total production time for items in the dataset.
}
\examples{
find_total_production_time(sample) # Calculate total production time for all items in the sample dataset
}
