\name{getPassword}
\alias{getPassword}
\title{Prompts the user ofr a password
http://stackoverflow.com/questions/2917202/r-way-to-securely-give-a-password-to-r-application}
\usage{
  getPassword()
}
\description{
  Prompts the user ofr a password
  http://stackoverflow.com/questions/2917202/r-way-to-securely-give-a-password-to-r-application
}

