% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{lexer}
\alias{lexer}
\title{Lexer for a makefile}
\usage{
lexer(makefile)
}
\arguments{
\item{makefile}{a string containing the relative path of the makefile}
}
\value{
A character vector of the tokens of the makefile.
}
\description{
Lexer for a makefile
}
\examples{
## NOT RUN
## tokens <- lexer("path/to/makefile")
}
