% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean.r
\name{make_clean}
\alias{make_clean}
\title{Clean function}
\usage{
make_clean(x)
}
\arguments{
\item{output}{Outputfile, typically `Makefile`}
}
\value{
`make_fun()` returns the input `x` invisibly.
}
\description{
Make Makefile of your R project
}
\examples{
#tmp <- tempdir()
#ex_dir <- file.path(system.file("simple", package = "MakeR"), "")
#system(paste("tree", ex_dir))
#make_fun(ex_dir, "Makefile")
#system(paste0("cat ", ex_dir, "/Makefile"))
#make_clean(paste0(ex_dir, "Makefile"))
}
\seealso{


Other make functions: 
\code{\link{make_fun}()}
}
\concept{make functions}
