% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_make.r
\name{make_fun}
\alias{make_fun}
\title{Make function}
\usage{
make_fun(path = NULL, output = "Makefile", dir_omit = NULL, clean = T)
}
\arguments{
\item{path}{Path to output file}

\item{output}{Outputfile, typically `Makefile`}

\item{clean}{Should return cleaned Makefile? (default = T)}
}
\value{
`make_fun()` returns the input `x` invisibly.
}
\description{
Make Makefile of your R project
}
\examples{
tmp <- tempdir()
ex_dir <- file.path(system.file("simple", package = "MakeR"), "")
system(paste("tree", ex_dir))
make_fun(ex_dir, "Makefile")
}
\seealso{


Other make functions: 
\code{\link{make_clean}()}
}
\concept{make functions}
