\docType{package}
\name{makeR-package}
\alias{makeR-package}
\title{Package for managing projects with multiple versions derived from a
       single source repository.}
\description{
  Package for managing projects with multiple versions
  derived from a single source repository.
}
\details{
  An R package to help manage R projects (e.g. Sweave
  reports) where multiple versions are created based upon a
  single source repository. For example, a monthly report
  where each versions is identitcal with the exception of
  easily definable parameters (e.g. date ranges for data
  extraction, title, etc.).
}
\author{
  \email{jason@bryer.org}
}
\keyword{Sweave}
\keyword{templating}
\keyword{versioning}

