% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeR.r
\docType{package}
\name{MakeR2-package}
\alias{MakeR2}
\alias{MakeR2-package}
\title{makeR: reproducible computation for R}
\description{
test
}
\details{
test2
}
\author{
Masatoshi Katabuchi <mattocci27@gmail.com>
}
\keyword{internal}
