% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maker.R
\name{make}
\alias{make}
\alias{maker}
\title{Transitional wrapper around remake}
\usage{
make(..., maker_file = "maker.yml")

maker(maker_file = "maker.yml", ...)
}
\arguments{
\item{...}{Arguments passed to remake}

\item{maker_file}{Passed to remake as remake_file}
}
\description{
Don't use this function; use the package remake
(https://github.com/richfitz/remake) instead.
}

