\name{parseProjectXML}
\alias{parseProjectXML}
\title{This is an internal method and should not be called directly.}
\usage{
  parseProjectXML(projectDir = getwd(),
    filename = "PROJECT.xml")
}
\arguments{
  \item{projectDir}{the directory containing the project
  file.}

  \item{filename}{the name of the project file.}
}
\description{
  This function will parse the XML project file from the
  given directory.
}

