\name{print.Project}
\alias{print.Project}
\title{Generic S3 method to print summary information about a Project class.}
\usage{
  \method{print}{Project} (x, ...)
}
\arguments{
  \item{x}{the Project}

  \item{...}{other unspecified parameters.}
}
\description{
  Generic S3 method to print summary information about a
  Project class.
}

