\name{print.Version}
\alias{print.Version}
\title{Generic S3 method for printing information about a Version class.}
\usage{
  \method{print}{Version} (x, ...)
}
\arguments{
  \item{x}{the Version}

  \item{...}{other non-specified parameters}
}
\description{
  Generic S3 method for printing information about a
  Version class.
}

