\name{releaseVersion}
\alias{releaseVersion}
\title{This is an internal method and should not be called directly.}
\usage{
  releaseVersion(pv, major = NULL, increment = TRUE, ...)
}
\arguments{
  \item{pv}{the Project}

  \item{major}{the major version number or name that should
  be released.}

  \item{increment}{should the minor version number be
  incremented.}

  \item{...}{other non-specified parameters}
}
\description{
  Release the latest built version.
}

