\name{setAutoOpen}
\alias{setAutoOpen}
\title{Sets whether built or released files will be opened automatically using the
system's default PDF viewer.}
\usage{
  setAutoOpen(value)
}
\arguments{
  \item{value}{if TRUE PDF files will be opened
  automatically.}
}
\description{
  Sets whether built or released files will be opened
  automatically using the system's default PDF viewer.
}

