\name{setAutoSave}
\alias{setAutoSave}
\title{Sets whether the PROJECT.xml file is automatically saved when project properties
have changed.}
\usage{
  setAutoSave(value)
}
\arguments{
  \item{value}{if TRUE PROJECT.xml will be saved
  automitcally.}
}
\description{
  Sets whether the PROJECT.xml file is automatically saved
  when project properties have changed.
}
\seealso{
  \code{Project$save()}
}

