\name{setDefaultBuilder}
\alias{setDefaultBuilder}
\title{Sets the default builder if none is specified to \code{Project$build}.}
\usage{
  setDefaultBuilder(value)
}
\arguments{
  \item{value}{the function to use as the default builder.}
}
\description{
  Sets the default builder if none is specified to
  \code{Project$build}.
}
\seealso{
  \code{\link{Project}}, \code{\link{builder.rnw}},
  \code{\link{builder.tex}},
  \code{\link{builder.cacheSweave}},
  \code{\link{builder.knitr}}
}

