% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis_fun}
\alias{vis_fun}
\title{vis function}
\usage{
vis_fun(make_dat, direction = "LR")
}
\arguments{
\item{output}{Outputfile, typically `Makefile`}
}
\value{
`make_fun()` returns the input `x` invisibly.
}
\description{
Make Makefile of your R project
}
\examples{
tmp <- tempdir()
ex_dir <- file.path(system.file("simple", package = "MakeR"), "")
system(paste("tree", ex_dir))
make_fun(ex_dir, "Makefile", clean = FALSE)
plan <- make_dat_fun(paste0(ex_dir, "Makefile"))
vis_fun(plan)
}
\seealso{

}
\concept{vis functions}
