\name{builder.tex}
\alias{builder.tex}
\title{This function will build LaTeX (tex) files using texti2pdf.}
\usage{
  builder.tex(sourceFile, theenv, ...)
}
\arguments{
  \item{sourceFile}{the source file to build.}

  \item{theenv}{the environment to build in.}

  \item{...}{other unspecified parameters}
}
\value{
  the name of the file if successfully built.
}
\description{
  This function will build LaTeX (tex) files using
  texti2pdf.
}

