% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis_fun}
\alias{vis_fun}
\alias{read_make}
\title{vis function}
\usage{
vis_fun(plan, direction = "LR")

read_make(x)
}
\arguments{
\item{x}{path}

\item{output}{Outputfile, typically `Makefile`}
}
\value{
`make_fun()` returns the input `x` invisibly.
}
\description{
Make Makefile of your R project
}
\examples{
tmp <- tempdir()
ex_dir <- file.path(system.file("simple", package = "MakeR2"), "")
system(paste("ls", ex_dir))
make_fun(ex_dir, clean = FALSE)
plan <- read_make(paste0(ex_dir, "Makefile"))
vis_fun(plan)
}
\concept{vis functions}
